refactor: redesign source code structure
