+++ src/st-mode/exec.h
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#pragma once
+
+double execute_st_computation(struct img_dim *dim, struct img_spec *img_spec, struct p_args *args, void *filters);
