+++ src/utils/threads-general.c
@@ -1,88 +1,42 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#include "mt-utils.h"
+#include "../../libbmp/libbmp.h"
+#include "../../logger/log.h"
+#include "threads-general.h"
 #include "utils.h"
-#include <pthread.h>
-#include <stdint.h>
+#include "args-parse.h"
+#include "filters.h"
 #include <string.h>
+#include <math.h>
 
-uint8_t process_by_row(struct thread_spec *th_spec, uint16_t *next_x_block, uint16_t block_size, pthread_mutex_t *x_block_mutex)
+struct img_dim *init_dimensions(uint16_t width, uint16_t height)
 {
-	pthread_mutex_lock(x_block_mutex);
-	// printf("next_block: %u, height: %d\n", *next_x_block,
-	// 	th_spec->dim->height);
-
-	if (*next_x_block >= th_spec->dim->height) {
-		pthread_mutex_unlock(x_block_mutex);
-		th_spec->start_row = th_spec->end_row = 0;
-		return 1;
-	}
-	th_spec->start_row = *next_x_block;
-	*next_x_block += block_size;
-	pthread_mutex_unlock(x_block_mutex);
-
-	th_spec->start_column = 0;
-	th_spec->end_row = fmin(th_spec->start_row + block_size, th_spec->dim->height);
-	th_spec->end_column = th_spec->dim->width;
-
-	return 0;
+	struct img_dim *dim = malloc(sizeof(struct img_dim));
+	if (!dim)
+		return NULL;
+	dim->width = width;
+	dim->height = height;
+	log_debug("Width: %d, Height: %d\n", width, height);
+	return dim;
 }
 
-uint8_t process_by_column(struct thread_spec *th_spec, uint16_t *next_y_block, uint16_t block_size, pthread_mutex_t *y_block_mutex)
+struct img_spec *init_img_spec(bmp_img *input, bmp_img *output)
 {
-	pthread_mutex_lock(y_block_mutex);
-	//	printf("next_block: %d, width: %d\n", *next_y_block,
-	// th_spec->dim->width);
-
-	if (*next_y_block >= th_spec->dim->width) {
-		pthread_mutex_unlock(y_block_mutex);
-		th_spec->start_column = th_spec->end_column = 0;
-		return 1;
-	}
-
-	th_spec->start_column = *next_y_block;
-	*next_y_block += block_size;
-	pthread_mutex_unlock(y_block_mutex);
-
-	th_spec->start_row = 0;
-	th_spec->end_row = th_spec->dim->height;
-	th_spec->end_column = fmin(th_spec->start_column + block_size, th_spec->dim->width);
-
-	return 0;
+	struct img_spec *spec = malloc(sizeof(struct img_spec));
+	if (!spec)
+		return NULL;
+	spec->input_img = input;
+	spec->output_img = output;
+	return spec;
 }
 
-uint8_t process_by_grid(struct thread_spec *th_spec, uint16_t *next_x_block, uint16_t *next_y_block, uint16_t block_size, pthread_mutex_t *xy_block_mutex)
+void *init_thread_spec(struct p_args *args, struct filter_mix *filters)
 {
-	pthread_mutex_lock(xy_block_mutex);
-	if (*next_x_block >= th_spec->dim->height || *next_y_block >= th_spec->dim->width) {
-		pthread_mutex_unlock(xy_block_mutex);
-		th_spec->start_row = th_spec->end_row = 0;
-		th_spec->start_column = th_spec->end_column = 0;
-		return 1;
-	}
+	struct thread_spec *th_spec = malloc(sizeof(struct thread_spec));
+	if (!th_spec)
+		return NULL;
 
-	th_spec->start_row = *next_x_block;
-	th_spec->start_column = *next_y_block;
-	*next_y_block += block_size;
-
-	if (*next_y_block >= th_spec->dim->width) {
-		*next_y_block = 0;
-		*next_x_block += block_size;
-	}
-	pthread_mutex_unlock(xy_block_mutex);
-
-	th_spec->end_row = fmin(th_spec->start_row + block_size, th_spec->dim->height);
-	th_spec->end_column = fmin(th_spec->start_column + block_size, th_spec->dim->width);
-	// printf("Row: st: %d, end: %d, Column: st: %d, end: %d \n",
-	// th_spec->start_row, th_spec->end_row, th_spec->start_column,
-	// th_spec->end_column);
-
-	return 0;
-}
-
-uint8_t process_by_pixel(struct thread_spec *th_spec, uint16_t *next_x_block, uint16_t *next_y_block, pthread_mutex_t *xy_block_mutex)
-{
-	return process_by_grid(th_spec, next_x_block, next_y_block, 1, xy_block_mutex);
+	return th_spec;
 }
 
 void apply_filter(struct thread_spec *spec, struct filter cfilter)
@@ -184,34 +138,4 @@ void filter_part_computation(struct thread_spec *spec, char *filter_type, struct
 	} else {
 		fprintf(stderr, "Error: Wrong filter type parameter '%s'\n", filter_type);
 	}
-}
-
-struct img_dim *init_dimensions(uint16_t width, uint16_t height)
-{
-	struct img_dim *dim = malloc(sizeof(struct img_dim));
-	if (!dim)
-		return NULL;
-	dim->width = width;
-	dim->height = height;
-	//	printf("Width: %d, Height: %d\n", width, height);
-	return dim;
-}
-
-struct img_spec *init_img_spec(bmp_img *input, bmp_img *output)
-{
-	struct img_spec *spec = malloc(sizeof(struct img_spec));
-	if (!spec)
-		return NULL;
-	spec->input_img = input;
-	spec->output_img = output;
-	return spec;
-}
-
-void *thread_spec_init(void)
-{
-	struct thread_spec *th_spec = malloc(sizeof(struct thread_spec));
-	if (!th_spec)
-		return NULL;
-
-	return th_spec;
 }
